+++ configuration/src/secrets.rs
@@ -96,13 +96,8 @@ impl AgentSecrets {
                         network,
                     );
                 }
-                SignerConf::Aws { id, region } => {
+                SignerConf::Aws { id } => {
                     eyre::ensure!(!id.is_empty(), "ID for {} aws signer key empty!", network,);
-                    eyre::ensure!(
-                        !region.is_empty(),
-                        "Region for {} aws signer key empty!",
-                        network,
-                    );
                 }
                 SignerConf::Node => (),
             }